1.修复基本的BUG 2.完成大部分CRUD操作 3.简化开发

+++ src/main/java/com/battcn/platform/service/BaseService.java
@@ -39,10 +39,11 @@ public abstract class BaseService<T>
 
 	public void batchDeleteByPrimaryKey(Object ids[])
 	{
-		for (int i = 0; i < ids.length; i++)
-		{
-			deleteByPrimaryKey(ids[i]);
-		}
+		if (ids != null && ids.length > 0)
+			for (int i = 0; i < ids.length; i++)
+			{
+				deleteByPrimaryKey(ids[i]);
+			}
 	}
 
 	public T selectByPrimaryKey(Object key)
